+++ ui-test/tests/Icons.jsx
@@ -20,7 +20,7 @@ const TestIcons = React.createClass({
                 <CardContent>
                     <h5>Basic Icons with Fallback tooltips</h5>
                     <Icon name="edit"/>
-                    <Icon name="delete"/>
+                    <Icon name="remove"/>
                     <Icon name="arrow_firstpage"/>
                     <Icon name="arrow_prevpage"/>
                     <Icon name="arrow_nextpage"/>
@@ -33,6 +33,7 @@ const TestIcons = React.createClass({
                     <Icon name="sort"/>
                     <Icon name="hide"/>
                     <Icon name="access_forbidden"/>
+                    <Icon name="delete"/>
                     <h5>Other Icons</h5>
                     <Icon name="widgets" tooltip="no tooltip on the next (canonical) icon"/>
                     <Icon name="access_forbidden" tooltip={false}/>